+++ packages/core/src/devTools.ts
@@ -2,25 +2,45 @@ import { Interpreter } from '.';
 import { IS_PRODUCTION } from './environment';
 
 type AnyInterpreter = Interpreter<any, any, any>;
-interface DevInterface {
-  services: Set<AnyInterpreter>;
-  register(service: AnyInterpreter): void;
-  onRegister(listener: ServicesListener): void;
+type ServiceListener = (service: AnyInterpreter) => void;
+
+export interface XStateDevInterface {
+  register: (service: Interpreter<any>) => void;
+  unregister: (service: Interpreter<any>) => void;
+  onRegister: (
+    listener: ServiceListener
+  ) => {
+    unsubscribe: () => void;
+  };
+  services: Set<Interpreter<any>>;
 }
 
-type ServicesListener = (service: Set<AnyInterpreter>) => void;
+// From https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/globalThis
+export function getGlobal() {
+  if (typeof self !== 'undefined') {
+    return self;
+  }
+  if (typeof window !== 'undefined') {
+    return window;
+  }
+  if (typeof global !== 'undefined') {
+    return global;
+  }
 
-function getDevTools(): DevInterface | undefined {
-  const w = window as Window & { __xstate__?: DevInterface };
-  if (!!w.__xstate__) {
-    return w.__xstate__;
+  return undefined;
+}
+
+function getDevTools(): XStateDevInterface | undefined {
+  const global = getGlobal();
+  if (global && '__xstate__' in global) {
+    return (global as any).__xstate__;
   }
 
   return undefined;
 }
 
 export function registerService(service: AnyInterpreter) {
-  if (IS_PRODUCTION || typeof window === 'undefined') {
+  if (IS_PRODUCTION || !getGlobal()) {
     return;
   }